// src/database/enums/DatabaseTable.ts
export enum DatabaseTable {
  BOOKS_INFO = 'books-info',
  BOOK_GENDERS = 'book-genders',
  BOOK_CHAPTERS = 'book-chapters',
  BOOK_CHAPTER_OPTIONS = 'book-chapters-options',
  BOOK_GENDER_BY_BOOK_INFO = 'book-gender-by-book-info',
  BOOK_USER_STATUS_BY_BOOK_INFO = 'book-user-status-by-book-info',
  BOOK_CHAPTER_HISTORY = 'book-chapter-history',
}